refactor(dao): fold complex2 back into complex
`complex2` existed for one reason: its impls imported `HolderDao` directly,
so dishka could analyse them, while `complex/*` had to quote the annotation
under `TYPE_CHECKING` to break the cycle the holder's properties created.
That cycle is gone, every `complex/*` impl imports `HolderDao` natively, and
the two packages no longer differ in any way — not in the protocols they
implement (`complex/` already hosts `core.<area>.adapters` impls, waiver ones
among them), not in build tooling, not in any documented rule.

So the seven poll and admin waiver adapters join `WaiverApproverImpl` in
`complex/waiver.py`, and `complex2/` goes. No behaviour changes: the classes
are unchanged and `ComplexDaoProvider` / `WaiverProvider` / `AdminProvider`
keep providing exactly what they did, from the new path.

`WaiverApproverImpl` drops its redundant `WaiverVoteGetter` base while it is
here — `WaiverApprover` already extends it — and takes `Iterable` from
`collections.abc` like the classes it now sits with.
