+++ shvatka/infrastructure/db/dao/complex/waiver.py
@@ -1,16 +1,23 @@
-import typing
-from collections.abc import Sequence
+from collections.abc import Iterable, Sequence
 from dataclasses import dataclass
 
 from shvatka.core.interfaces.dal.waiver import WaiverApprover
 from shvatka.core.models import dto
 from shvatka.core.models.enums import Played
-from shvatka.core.waiver.adapters import WaiverVoteGetter
+from shvatka.core.waiver.adapters import (
+    AdminGameWaiversReader,
+    AdminPollReader,
+    AdminWaiverEditor,
+    PollDraftsReader,
+    PollVoteRemover,
+    WaiverVoteAdder,
+    WaiverVoteGetter,
+)
 from shvatka.infrastructure.db.dao.holder import HolderDao
 
 
 @dataclass
-class WaiverApproverImpl(WaiverApprover, WaiverVoteGetter):
+class WaiverApproverImpl(WaiverApprover):
     dao: HolderDao
 
     async def upsert(self, waiver: dto.Waiver) -> None:
@@ -37,7 +44,131 @@ class WaiverApproverImpl(WaiverApprover, WaiverVoteGetter):
     async def get_dict_player_vote(self, team_id: int) -> dict[int, Played]:
         return await self.dao.poll.get_dict_player_vote(team_id)
 
-    async def get_by_ids_with_user_and_pit(
-        self, ids: typing.Iterable[int]
-    ) -> list[dto.VotedPlayer]:
+    async def get_by_ids_with_user_and_pit(self, ids: Iterable[int]) -> list[dto.VotedPlayer]:
         return await self.dao.player.get_by_ids_with_user_and_pit(ids)
+
+
+@dataclass
+class WaiverVoteAdderImpl(WaiverVoteAdder):
+    dao: HolderDao
+
+    async def is_excluded(
+        self,
+        game: dto.Game,
+        player: dto.Player,
+        team: dto.Team,
+    ) -> bool:
+        return await self.dao.waiver.is_excluded(game, player, team)
+
+    async def add_player_vote(self, team_id: int, player_id: int, vote_var: str) -> None:
+        return await self.dao.poll.add_player_vote(team_id, player_id, vote_var)
+
+    async def get_team_player(self, player: dto.Player) -> dto.TeamPlayer:
+        return await self.dao.team_player.get_team_player(player)
+
+
+@dataclass
+class WaiverVoteGetterImpl(WaiverVoteGetter):
+    dao: HolderDao
+
+    async def get_dict_player_vote(self, team_id: int) -> dict[int, Played]:
+        return await self.dao.poll.get_dict_player_vote(team_id)
+
+    async def get_by_ids_with_user_and_pit(self, ids: Iterable[int]) -> list[dto.VotedPlayer]:
+        return await self.dao.player.get_by_ids_with_user_and_pit(ids)
+
+
+@dataclass
+class PollDraftsReaderImpl(PollDraftsReader):
+    dao: HolderDao
+
+    async def get_polled_teams(self) -> list[int]:
+        return await self.dao.poll.get_polled_teams()
+
+    async def get_by_ids_with_user_and_pit(self, ids: Iterable[int]) -> list[dto.VotedPlayer]:
+        return await self.dao.player.get_by_ids_with_user_and_pit(ids)
+
+    async def get_dict_player_vote(self, team_id: int) -> dict[int, Played]:
+        return await self.dao.poll.get_dict_player_vote(team_id)
+
+    async def get_by_id(self, id_: int) -> dto.Team:
+        return await self.dao.team.get_by_id(id_)
+
+    async def get_by_player_or_none(
+        self, game: dto.Game, player: dto.Player
+    ) -> dto.SecondaryOrganizer | None:
+        return await self.dao.organizer.get_by_player_or_none(game=game, player=player)
+
+    async def get_by_player(self, game: dto.Game, player: dto.Player) -> dto.SecondaryOrganizer:
+        return await self.dao.organizer.get_by_player(game=game, player=player)
+
+
+@dataclass
+class AdminPollReaderImpl(AdminPollReader):
+    dao: HolderDao
+
+    async def get_polled_teams(self) -> list[int]:
+        return await self.dao.poll.get_polled_teams()
+
+    async def get_by_ids_with_user_and_pit(self, ids: Iterable[int]) -> list[dto.VotedPlayer]:
+        return await self.dao.player.get_by_ids_with_user_and_pit(ids)
+
+    async def get_dict_player_vote(self, team_id: int) -> dict[int, Played]:
+        return await self.dao.poll.get_dict_player_vote(team_id)
+
+    async def get_by_id(self, id_: int) -> dto.Team:
+        return await self.dao.team.get_by_id(id_)
+
+
+@dataclass
+class PollVoteRemoverImpl(PollVoteRemover):
+    dao: HolderDao
+
+    async def del_player_vote(self, team_id: int, player_id: int) -> None:
+        return await self.dao.poll.del_player_vote(team_id, player_id)
+
+
+@dataclass
+class AdminGameWaiversReaderImpl(AdminGameWaiversReader):
+    dao: HolderDao
+
+    async def get_by_id(self, id_: int) -> dto.Game:
+        return await self.dao.game.get_by_id(id_)
+
+    async def get_played_teams(self, game: dto.Game) -> Iterable[dto.Team]:
+        return await self.dao.waiver.get_played_teams(game)
+
+    async def get_played(self, game: dto.Game, team: dto.Team) -> Iterable[dto.VotedPlayer]:
+        return await self.dao.waiver.get_played(game, team)
+
+    async def get_all_by_game(self, game: dto.Game) -> list[dto.Waiver]:
+        return await self.dao.waiver.get_all_by_game(game)
+
+
+@dataclass
+class AdminWaiverEditorImpl(AdminWaiverEditor):
+    dao: HolderDao
+
+    async def get_game_by_id(self, id_: int) -> dto.Game:
+        return await self.dao.game.get_by_id(id_)
+
+    async def get_team_by_id(self, id_: int) -> dto.Team:
+        return await self.dao.team.get_by_id(id_)
+
+    async def get_player_by_id(self, id_: int) -> dto.Player:
+        return await self.dao.player.get_by_id(id_)
+
+    async def get_team_player(self, player: dto.Player) -> dto.TeamPlayer:
+        return await self.dao.team_player.get_team_player(player)
+
+    async def get_team_waivers(self, game: dto.Game, team: dto.Team) -> list[dto.Waiver]:
+        return await self.dao.waiver.get_all_by_team(game, team)
+
+    async def upsert(self, waiver: dto.Waiver) -> None:
+        await self.dao.waiver.upsert(waiver)
+
+    async def delete(self, waiver: dto.WaiverQuery) -> None:
+        await self.dao.waiver.delete(waiver)
+
+    async def commit(self) -> None:
+        await self.dao.commit()
